Added MVC Architecture
