+++ models/UserModel.js
@@ -1,0 +1,12 @@
+import mongoose from "mongoose";
+const UserSchema = mongoose.Schema({
+  email: {
+    type: String,
+    required: true,
+  },
+  password: {
+    type: String,
+    required: true,
+  },
+});
+export default UserSchema;
