+++ routes/authRoutes.js
@@ -1,0 +1,4 @@
+import express from "express";
+const router = express.Router();
+router.post("/signup");
+export default router;
